Fixed issues with creating the CalendarEvent table
