+++ .gitignore
@@ -34,6 +34,7 @@
 
 # env files (can opt-in for committing if needed)
 .env*
+env
 
 # vercel
 .vercel
