+++ backend/Entities/User.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace StudyVerseBackend.Entities;
 
-public class User: IdentityUser
+public class User : IdentityUser
 {
     public string Name { get; set; } = string.Empty;
     [Required]
@@ -14,8 +14,10 @@ public class User: IdentityUser
 
     [Required, EmailAddress] public override string Email { get; set; }
     public string Avatar_Url { get; set; } = string.Empty;
-    
+
     public string? CustomizationOptions { get; set; }
+
+    public ICollection<CalendarEvent> CalendarEvents { get; } = new List<CalendarEvent>();
 
     public void SetCustomizationSettings(string settings)
     {
