+++ backend/Infastructure/Contexts/ApplicationDbContext.cs
@@ -10,13 +10,10 @@ public class ApplicationDbContext
  * This is the class where the Database context will reside, which is responsible for representing the
  * Tables as Models in C#.
  */
-
-    public DbSet<User> Users { get; set; }
-
-
-
-
 {
+    public DbSet<CalendarEvent> CalendarEvents { get; set; }
+
+
     public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
     {
     }
@@ -25,6 +22,7 @@ public class ApplicationDbContext
     {
         base.OnModelCreating(builder);
 
+        // Configuring relationships with users
         builder.Entity<User>()
             .HasIndex(u => u.UserName)
             .IsUnique();
@@ -36,6 +34,16 @@ public class ApplicationDbContext
         builder.Entity<User>()
             .Property(u => u.CustomizationOptions)
             .HasColumnType("jsonb");
+
+        /*
+         * This code creates the following relationship:
+         * A user can have many events, but a event must HAVE one user
+         */
+        builder.Entity<User>()
+            .HasMany(e => e.CalendarEvents)
+            .WithOne(e => e.User)
+            .HasForeignKey(e => e.UserId)
+            .IsRequired();
     }
 
 }
