--- env
@@ -1,7 +1,0 @@
-# DATABASE INFORMATION
-DB_CONNECTION_STRING=Host=db.ulgbaidfotsuuofflocd.supabase.co;Database=postgres;Username=postgres;Password= StudyV3rs3123;SSL Mode=Require;Trust Server Certificate=true
-
-# Authentication endpoint
-JWTCONFIG_VALID_AUDIENCE=http://localhost:5149
-JWTCONFIG_VALID_ISSUER=http://localhost:5149
-JWTCONFIG_SECRET=db7c8d3ab109355d6ca265d5c29bc38a72a760e84b75942dfd4b61f738e5eda77be07fc71483103d450dd429cae37287fb6f43a68fa6d1172a55ae10bf1509c69a9f2c810327ff4f257890d8a31e56d71688766d6ac69555f0ccfa43d3203a2f06329f62fab5db0eaefb5cc8f58a545547ad10c5f208d2e40d8e91d7092000015835723705ae12540a98028218a9c4e33f1be1f3bf93858a23d289ed522e3a5d3f2a2d46423d589b59a5e07313ac5c50dad6fb1a9478e6c56c6e16d6d8c54494bb88f25573abbcfab9d404c93dc38f97ff49dc5cf72d05abee8327fc588a87fae82bad11e1aae8fe27b0cc606d46a661007d5b5abd9119790800ef4ad6640901
